adds auth_config to utils folder, moves utils to src not components
