+++ Front-end/client/src/components/App/Nav/Auth0Nav.js
@@ -1,7 +1,7 @@
 //component renders two buttons, for logging in and logging out, depending on whether the user is currently authenticated
 
 import React from "react";
-import { useAuth0 } from "../utils/react-auth0-spa";
+import { useAuth0 } from "../../../utils/react-auth0-spa";
 
 const NavBar = () => {
   const { isAuthenticated, loginWithRedirect, logout } = useAuth0();
