+++ Front-end/client/src/index.js
@@ -4,8 +4,8 @@ import "./index.css";
 import App from "./components/App/App";
 import registerServiceWorker from "./registerServiceWorker";
 
-import { Auth0Provider } from "./react-auth0-spa";
-import config from "./auth_config.json";
+import { Auth0Provider } from "./utils/react-auth0-spa";
+import config from "./utils/auth_config.json";
 import history from "./utils/history";
 
 // A function that routes the user to the right place
