+++ Front-end/client/src/utils/auth_config.json
@@ -1,0 +1,4 @@
+{
+  "domain": "dev-lnof6z3s.auth0.com",
+  "clientId": "WfHlntNa4YtEEgXdsd43La5sehnar5DV"
+}
